requirements.txt
python-dotenv
tweepy
kafka-python